[refactor] emoji, comment, folder를 분리 #4
