--- src/main/kotlin/dev/codeswamp/domain/article/entity/rdb/article/ArticleType.kt
@@ -1,6 +1,0 @@
-package dev.codeswamp.domain.article.entity.rdb.article
-
-enum class ArticleType {
-    DRAFT,
-    PUBLISHED,
-}
