+++ src/main/kotlin/dev/codeswamp/domain/folder/controller/FolderController.kt
@@ -1,0 +1,8 @@
+package dev.codeswamp.domain.folder.controller
+
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class FolderController {
+
+}
